Install Atuin
[Atuin][1] looks interesting: it puts your command line history into
a database (and allows you to sync this between machines). Let's give it
a try...

At the moment I can't install it on my home machine (too old), but let's
try it on my work laptop...

As well as installing Atuin itself, let's import the existing history
(is this idempotent? Not sure, but hopefully the worst that could happen
is some duplication) and also generate completions for it.

[1]: https://atuin.sh