.gitconfig: use "histogram" diff algorithm
"histogram" is an alternative diff algorithm which seems to do
slightly better in some contexts at identifying code that has moved
or been rearranged.

I enabled it a while ago and to the extent that I notice any
difference, I prefer the "histogram" version of the diff, so let's
enable it officially.