Bump asdf version and add "asdf update" task
`asdf` has its own `asdf update` command, and since the standard way
to install it is via git, this just fetches and switches to the
latest release.

Since this is handy to know about, let's add it here as a reminder.
This does mean that the `update` task will always return a "changed"
status at the moment, but I don't think that matters too much.

For the most part, it's probably fine to just be on the latest version
of asdf. I'll continue specifying the version for now mainly just
because that avoids the need to make any changes to the install step.