Fix linting issues with bootstrap script
I stumbled across [shellcheck][1], a sort of linter for shell scripts,
and ran this script through it to see what it said.

It suggested that backticks are deprecated and `$()` is preferable.
It also suggested wrapping args in double quotes, in order to avoid
"globbing". That certainly sounds like something I would like to avoid.

[1]: https://www.shellcheck.net