Manage Atuin's config in .dotfiles
This commit copies the default config from its default location into my
dotfiles, and adds a line to .bashrc to point to it.

Note that because Atuin only seems to let you specify a config
directory, not a config file, I've put the generically-named
"config.toml" into an `atuin` folder.