Fix issue where Ansible tags weren't working as intended
Ansible started complaining that `include` was deprecated.
In an [earlier commit][1] I updated these to `include_tasks` instead,
which fixed the deprecation warning.

However, I now discover that it broke the ability to target tasks by
tag. I'm not completely clear why this is (Ansible's docs are a bit
baffling to me), but switching from `include_tasks` to `import_tasks`
seems to fix this issue.

For future reference: you can use the `--list-tasks` flag to see what
tasks will be run for a given tag. For example:

    ANSIBLE_CONFIG=ansible/ansible.cfg ansible-playbook -v -K -i ansible/hosts ansible/playbook.yml --list-tasks --tags "apt"

...will list the tasks that are tagged "apt".

[1]: fe9ee72