Add a default blame.ignoreRevsFile
`blame.ignoreRevsFile` allows you to specify uninteresting commits in a
file, so that `git blame` will ignore these commits. This is useful for,
eg. a commit that applied Prettier to thousands of files, but didn't
make any meaningful change to the code.

Let's set a default filename for this so I can use it without having to
specify it for each repo.