Disable corepack auto-pinning
Corepack has recently started automatically adding a `packageManager`
field to `package.json` if it's not already present.

This is quite annoying when working on projects where I don't want to
add this field, which is currently a lot of projects. I mainly want to
use corepack as a generic way of managing `yarn` / `pnpm` versions,
with the *option* to enforce it via a `packageManager` field on my
own projects.

Setting `COREPACK_ENABLE_AUTO_PIN` to `0` seems to fix this issue [1],
so let's do that.

[1]: https://github.com/nodejs/corepack?tab=readme-ov-file#environment-variables