Change git diff.colorMoved to fix issue with delta diffs
The value `plain` that I was using doesn't seem to work so nicely with
`delta` - it causes odd striping in some files (eg. a file that was
entirely new looks as if some lines are not new).

Changing `diff.colorMoved` to `default` [1] fixes this and moved blocks
still look as they should.

[1]: https://git-scm.com/docs/git-diff#Documentation/git-diff.txt---color-movedltmodegt