Revert "Add a default blame.ignoreRevsFile"
It turns out that setting this config setting causes `git blame` to fail
if the specified file doesn't exist. I thought I might be able to
side-step this by creating an empty `.git-blame-ignore-revs` file in my
home directory but that doesn't work either.

So it seems that for now at least it's better to set
`blame.ignoreRevsFile` on a per-project basis rather than globally.
Oh well.

This reverts commit 3ca6567b976d8dcb8b5c36ebda0d3a31f1bb8fe5.