Install catpuccin theme for nvim
Let's see if we can make vim look a bit nicer...

This also adds the [vim-plug][1] plugin manager to install [catpuccin][2]
I don't know whether this is the best or standard option but I picked it
because a) it seemed simple-ish to install, and b) it's by the same
person who made fzf, which I like.

I've set `transparent_background` to `true` to avoid it looking odd when
catpuccin's background colour doesn't match the terminal's.

[1]: https://github.com/junegunn/vim-plug
[2]: https://github.com/catppuccin/nvim