Use latest available version of Neovim
My work machine is on a slightly older version of Ubuntu, whose Neovim
is on a version that is too old to work with Catppuccin (the theme I'm
using).

Using the latest unstable release of Neovim might not be the *best* way
to sidestep this problem, but it is the quickest and easiest and works
for now.

nb. You might still need to run `:PlugInstall` when launching nvim for
the first time in order to install Catppuccin.