Fix colour contrast issue with ls
Directories with broad write permissions have blue text on a green
background when I use `ls` with a colourised output. This is fairly
unreadable. This tweak adjusts the text colour in this context to dark
grey, making it a lot more readable.

See this Stack Exchange question for a little more info:
https://serverfault.com/q/183364/124678