Define a custom hunk header pattern for JavaScript diffs
Git has a bunch of built-in patterns for various languages to give
context to the hunks within a diff. For some reason it doesn't have a
pattern for JavaScript, but it's possible to [define custom patterns][1].

This is my attempt at defining one for JS. It tries to match the nearest
class, function, variable, or method declaration (by looking for the
keyword in a line that ends with a `{`). The last bit of the regex is a
bit funky: it's trying to match arbitrary methods on a class by looking
for anything followed by `()` and then `{` (eg. `render() {`)

This seems to work quite well! The resulting hunk headers seem to
generally catch the surrounding function, anyway. Maybe there's
something I'm overlooking, because it's puzzling to me that javascript
isn't included among the default patterns.

Oh, by the way, I've added this for TypeScript too, but I haven't
actually tested with TS - I'm just figuring that it'll *probably* be
better than nothing...

[1]: https://git-scm.com/docs/gitattributes#_defining_a_custom_hunk_header