delta: adjust line number colours to improve readability
The default colours for added / deleted line numbers are quite dark,
and not very readable. Setting them to brighter colours here makes them
easier to read.

Note that "tomato" is a CSS colour, whereas "green" is (I think) using
the terminal's definition of "green" (which is also a CSS colour, but
the ANSI version seems to take priority).