.bash_prompt: don't check for .git-coauthors file
An annoyance about having git mob's initials in my prompt is that if I'm
on a machine where `.git-coauthors` doesn't exist (eg. my home PC, where
I never need git mob), the attempt to add the initials causes an error
every time I run a command.

For this reason I was checking for the existence of this file. But I've
realised that this is responsible for a perceptible delay every time I
hit Enter.

We can fix the error message issue without needing to check the file
system each time by checking an environment variable instead. If the
`_DISABLE_GIT_MOB_INITIALS` is set (eg. in the .gitignored `.extra`
file) then we just won't even try to do anything with git mob.