Install winget apps via Ansible
winget is great for installing and updating apps in Windows. Ideally it
would be nice to automate installing stuff in Windows the exact same way
I install things in WSL.

Since winget can be run from WSL, we can write an Ansible task to run
`winget install` on a bunch of apps. This seems to work quite nicely,
though requires a little manual input in the form of responding to
confim prompts from Windows permission control. You could probably avoid
this by running it in a terminal with admin privileges.