Revert "tig: always split panels vertically"
A while ago I set tig to always split panels vertically. This seemed
good on my personal laptop, where the screen is *just* big enough for
the vertical split to always be comfortable.

In practice though, I find myself sometimes wishing tig was responsive
(when, eg. viewing it in a smaller window). So let's revert this change
for now and see how I get on with the defaults again.

This reverts commit a7cdfeed5723c6db1e2d5eb9003ab52874457680.