Add commit hash to output of "git br"
I just accidentally deleted the wrong branch. It didn't matter (it was a
branch I'd have deleted anyway), but it made me realise that it would be
useful to have the commit hash in the output of "git br" too: that would
make it easier to restore a deleted branch in this situation.

While doing this I'll change the authorname colour, just to keep them
all different colours.