Add "git bri" alias to quickly browse branches in tig
This is very similar to "git coi", only it passes the selected branch to
tig rather than checking it out.

Note that tig goes into pager mode when you pipe things into it. I
needed to add a `--stdin` flag in order to get it to behave as I
expected (ie. treat the branch name as an argument)

Fwiw, the rationale for the alias is "git branches interactive"