Install Helix via official PPA repository
Rather than download the latest release from Github, let's use the PPA
repository. This has the advantage of being simpler and more reliable,
but the disadvantage that we might be on a slightly older version.
I'm just trying it out so that's ok for now.

Keeping the previous commit rather than squashing it just in case I ever
want to use the latest version instead.