Helix: Add "jk" shortcut for exiting insert mode
This avoids the need to hit Escape to switch modes: if you type "j"
followed by "k" in insert mode you'll switch to normal mode. If you
follow "j" with any other letter, both letters will appear, so it
shouldn't mess with typing too much, unless you want to type "jk", as in
this commit message, which is slightly awkward. Fortunately that's not a
letter combination that comes up very often.