mise: Install completions in a way that actually works
For some reason (perhaps to do with how bash_completions is installed)
the suggested approach for doing this didn't work. When I `sudo` create
the completions they don't include the bash completions lib, and when
I try other ways to do it (create it with a non-sudo user and then copy
the file to `/etc/bash_completion.d/mise`) new terminals just hang.

Instead, let's follow the same pattern as Atuin: creating the
completions in the home directory and then sourcing it directly in
`.bashrc`. This seems to work.