Install global tools via mise instead of npm
A minor annoyance I've had with mise is that when in a project with
a specified NodeJS version that is different from the global NodeJS
version you can no longer run globally-installed NodeJS tools: instead
I was having to reinstall them in each project.

This didn't happen with asdf, but mise seems a bit more strict
about not falling back to the global NodeJS version in this case.
Installing these NodeJS tools via mise seems to mostly get around
this, in that they no longer need to be reinstalled for each version
of NodeJS (presumably because mise now knows where they live + how to
shim them).

There are two minor issues to bear in mind:

- In a project that uses a version of NodeJS that's not available, you
  do still have to install NodeJS for that project (via `mise install`)
  to be able to use these tools. However, you don't need to reinstall
  the tools themselves.

- For upgrading npm itself, let's continue to use Ansible's built-in
  npm module. This is because `npm` is already available globally, and
  while `mise use --global npm:npm@latest` does seem to install
  the requested version, it installs it separately and mise continues
  to use the version that was installed with NodeJS (and I don't know
  how to tell it to do otherwise). Using npm directly upgrades the
  global version, which is what we want, and is also just less
  confusing all round. My impression is that at the moment mise
  [isn't capable of managing npm versions itself][1].

Another potentially good thing about installing global tools via mise
is that `mise ls` becomes a more comprehensive list of the tools
that are installed.

[1]: https://github.com/jdx/mise/discussions/4185#discussioncomment-11931397