Use Zellij
Adding this to .bashrc means we'll automatically start a Zellij session
(if Zellij is available) and also exit the shell when we exit.

I'm not sure I want this in VS Code's terminal, where space is already
limited (and I don't tend to use tabs anyway), so I've added an
environment variable that allows me to distinguish that context and not
use Zellij in that case.

The way I did this was by setting:

    "terminal.integrated.env.linux": {
      "VS_CODE_TERMINAL": "true",
    },

...in VS Code's user settings.