WIP: Add custom Starship config
The nice thing about Starship is that it is cross-platform and highly
customisable. My current prompt is a cobbled-together bash thing;
switching to Starship would mean a) it should "just work" on different
platforms, and b) it should be easier to customise.

This custom config more-or-less replicates my existing prompt in
Starship. The one problem is that git-mob is a bit slower...

Note that there's a little bit of fiddlyness around showing the close
bracket after git status: we want the space + status to be conditional
(wrapped in parentheses), but we always want the close bracket to
display. In order to achieve this we need to move the close bracket "\)"
outside the conditional, and then separately apply the style to it
(otherwise it would be the wrong colour).