Update Zellij config
Hmm, when I made a settings tweak in Zellij (stop showing hints on
startup) it completely rejigged this file, which makes me wonder if
having it in my dotfiles repo is ideal. But it's not clear to me how
I could otherwise have custom config (as far as I can tell KDL doesn't
have a concept of includes/imports).

For now let's just commit the updated version and see whether this is
something that happens often enough to be annoying.