Install Docker
This is pretty much just the instructions on [this page][1] as an
Ansible task to remind myself how I installed Docker.

IMPORTANT NOTE: on Linux Mint there seems to be a bug with Docker
Desktop where it decides to use all available space by default, and
ends up eating all the free space on the drive. In order to side-step
this I needed to allocate space in Docker's settings by adding a line
`"DiskSizeMiB": 64000` to `~/.docker/desktop/settings-store.json`!

Hat tip to [this post][2].

[1]: https://docs.docker.com/desktop/setup/install/linux/ubuntu/
[2]: https://forums.docker.com/t/docker-4-39-0-not-running-able-to-start-on-ubuntu-24-10/147171/4