Starship: do not truncate the path when we're in a repo
I was getting confused because I had the same project in two places.
To avoid this, it's useful to see the full path: ie. instead of
"my-project" we might see "/path/to/my-project"

Starship allows us to style the repo name differently, so let's make it
more prominent than the preceding path.

Let's also set the commit hash to the same style because it was a bit
too low contrast in some places.