Move all Windows-specific config into its own files
Some of these (in particular setting BROWSER in .bashrc) cause issues
in Linux. To avoid this, let's move the Windows-specific stuff into its
own place.

Note that this will need to be conditionally included to work on
Windows. I'll leave that as a problem for later though, because I'm not
currently on Windows.