Install Helix directly rather than from PPA
The PPA is stuck on an older version of Helix, and has now been
archived. To get the latest version, let's just install the .deb package
directly from Helix's releases page.

Because the numbering is slightly odd here, I've hardcoded the whole
URL. When there's a new version, update the URL.