Add Ghostty install task
Note that this is running a remote shell script, which isn't totally
ideal. It's worth checking the contents of this script before running
this. Perhaps I should consider moving the current version of the script
(which I've checked) into this repo... Anyway, I'm including this for
now just as a record of how I installed Ghostty.